Fix binding for "capitalize keywords" in sql-layer
Noticed a new warning during startup saying
that sqlup-mode is using "=c" where "=" is no
prefix key. Noticed that the entire binding
was not carried out in this case.

Therefore I have removed "=" from the binding
but left it on "c" until enough text
manipulations are available, then it should
be moved to "x text manipulation" group instead.